+++ src/server/src/projects/entities/project.entity.ts
@@ -58,4 +58,9 @@ export class Project implements IProject {
     default: "'{}'"
   })
   lockedDistricts: readonly boolean[];
+
+  // Strips out data that we don't want to have available in the read-only view in the UI
+  getReadOnlyView(): Project {
+    return { ...this, lockedDistricts: new Array(this.numberOfDistricts).fill(false) };
+  }
 }